Expose check_docker() test utility function
Move check_docker() from resolwe.flow.tests.test_executors to
resolwe.flow.utils.test and make it 'public'.
Use Sphinx's new ability to specify container types.
Use upper-case names for FLOW_* settings overrides.
