+++ docs/CHANGELOG.rst
@@ -32,6 +32,7 @@ Changed
 - Use Jinja2 instead of Django Template syntax
 - Expression engine must be declared in ``requirements``
 - Set Docker Compose's project name to ``resolwe`` to avoid name clashes
+- Expose ``check_docker()`` test utility function
 
 Fixed
 -----
